Fix #7 - Small screenshots (#8)
* Fix #7 - Small screenshots

This fixes issue #7, where screenshots would be very small when displayed for markup,
and continue shrinking every time you went back to edit the image.

The issue seems to be caused by an incorrect screen density being used when loading
the screenshot image. This causes the image to scale to a smaller image.

* Address Review Comments

Use full variable names, reference AOSP issue.