Add API for manually triggering the feedback flow
Consumers may want to create their own UI for starting
the feedback flow, without requiring the user to shake
their device. With this change, we now expose the
Shaky class to users of the library, but with only one
public method, startFeedbackFlow().

This also updates the demo app to show how to use the
new API.

Fixes #6