Add support to send feedback title for bottomsheet flow
Summary: In bottomsheet flow, users can choose between `Report a bug` and `Send app feedback`. Per previous flow it was mandatory to send title that would identify which option was being selected in the bottomsheet. Currently, only attachments are bring sent in `safeResult` and title field is empty.

Support Added:
- Added support `BottomSheetFeedbackItem#getTitleResId` to get title based on `BottomSheetFeedbackItem` feedback type.
- Added a new key `BottomSheetFeedbackTypeAdapter#EXTRA_FEEDBACK_TITLE` to pass the value of title in intent using the newly added method.
- Added support to fetch the value of feedback title when bottomsheet flow is active upon receiving the broadcast in `Shaky#createReceiver`.
- Further, support added to set value of title in `safeResult` to submit it in feedback.
- Also, removed the key ‘BottomSheetFeedbackTypeAdapter#EXTRA_FEEDBACK_TYPE’ as it has no usage in the bottomsheet flow.